% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{drop.avg}
\alias{drop.avg}
\title{Average grade with lowest scores dropped}
\usage{
drop.avg(x, drop)
}
\arguments{
\item{x}{A vector of grades.}

\item{drop}{Number of lowest grades to drop.}
}
\value{
The average of \code{x} after setting missing values to zero and dropping the lowest \code{drop} observations.
}
\description{
This function calculates the mean of a vector of student grades after dropping a user-specified number of the lowest scores.
}
\examples{
grades <- c(85, 90, NA, 100, 75)
drop.avg(grades, 2)
}

