% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{mse}
\alias{mse}
\title{Calculate mean squared error}
\usage{
mse(x, truth = 0, ...)
}
\arguments{
\item{x}{A vector of observations.}

\item{truth}{The true parameter value. Defaults to zero.}
}
\value{
The mean squared error of \code{x} centered around \code{truth}.
}
\description{
This function calculates mean squared error centered around a user-specified true value, which defaults to zero.
}
\examples{
y <- rnorm(10)
mse(y)
mse(y, trim 0.05)
y <- c(y, NA)
mse(y, na.rm = TRUE)
}

