% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{premail}
\alias{premail}
\title{Extract prefixes from a vector of email addresses}
\usage{
premail(x)
}
\arguments{
\item{x}{A vector of text strings, each of which is an email address.}
}
\value{
The email prefixes, i.e. everything before the @ symbol.
}
\description{
This function extracts the prefixes, i.e. everything before the @ symbol, from a vector of email addresses.
}
\examples{
emails <- c("foo@bar.com", "bar@foo.com")
premail(emails)
}

