% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{rmse}
\alias{rmse}
\title{Calculate root mean squared error}
\usage{
rmse(x, truth = 0, ...)
}
\arguments{
\item{x}{A vector of observations.}

\item{truth}{The true parameter value. Defaults to zero.}
}
\value{
The root mean squared error of \code{x} centered around \code{truth}.
}
\description{
This function calculates root mean squared error centered around a user-specified true value, which defaults to zero.
}
\examples{
set.seed(1)
y <- rnorm(10)
rmse(y)
rmse(y, trim 0.05)
y <- c(y, NA)
rmse(y, na.rm = TRUE)
}

