% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{zscores}
\alias{zscores}
\title{Center and standardize a vector}
\usage{
zscores(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of observations}

\item{na.rm}{Remove missing values? Defaults to TRUE.}
}
\value{
The z-scores corresponding to \code{x}.
}
\description{
This function converts a vector of observations to z-scores, by subtracting the sample mean and dividing by the sample standard deviation.
}
\examples{
y <- rnorm(10, 5, 10)
zscores(y)
}

